Almost finished project
The time sensing unit measure is in millisecond instead of second. For
this reason the pollution power consumption is expressed in uJ.
The overall energy consumption is affected from these four factors:
1. Radio duty cycle (More RDC more E.C.)
2. Sampling time of the pollution data (More sampling time less is the
E.C.)
3. Sensing time. i.e. how much time the pollution sensors require to
take the values (More sensing time more is the E.C.)
4. the number of nodes in the WSN  (More number of nodes time more is
the E.C.)

E.C. is the energy consumption.

+++ NES_Proj/energy/energy_data_structure.h
@@ -2,7 +2,7 @@
 #define ENERGY_CONS_H 3424
 struct energy_cons{
 
-uint16_t time_sensing;
+unsigned long time_sensing;
 //cpu energy cons
 unsigned long cpu;
 unsigned long lpm;
